+++ example/example_task_cancel_during_stream_iteration.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
+import 'package:defer/defer.dart';
 import 'package:multitasking/multitasking.dart';
-import 'package:multitasking/stream/cancellable_stream_iterator.dart';
 
 Future<void> main(List<String> args) async {
   final controller = StreamController<int>.broadcast();
@@ -51,20 +51,20 @@ Task<int> _doWork(Stream<int> stream, CancellationToken token,
   return Task.run(() async {
     await Task.sleep();
     final list = <int>[];
-    final iterator = CancellableStreamIterator(stream, token);
-    try {
-      while (await iterator.moveNext()) {
-        final event = iterator.current;
-        _message('Received event: $event');
-        list.add(event);
-        if (list.length == 1 && testBreak) {
-          _message('I want to break free...');
-          break;
+    final iterator = StreamIterator(stream);
+    await runCancellable(token, iterator.cancel, () async {
+      await defer(iterator.cancel, () async {
+        while (await iterator.moveNext()) {
+          final event = iterator.current;
+          _message('Received event: $event');
+          list.add(event);
+          if (list.length == 1 && testBreak) {
+            _message('I want to break free...');
+            break;
+          }
         }
-      }
-    } finally {
-      await iterator.cancel();
-    }
+      });
+    });
 
     token.throwIfCancelled();
     await Task.sleep();