+++ lib/src/task/cancellation.dart
@@ -1,9 +1,71 @@
 import 'dart:async';
 
-import 'package:defer/defer.dart';
-import 'package:meta/meta.dart';
-
 import 'errors.dart';
+
+/// Performs the following actions:
+///
+/// - Creates a separate zone for catching exceptions
+/// - Executes the [action] function
+/// - Waits for the function [action] to complete and returns the result (value
+/// or error)
+/// - Ignores (suppresses) all unhandled exceptions in the created zone after
+/// returning the result
+///
+/// This method can be used for the case where a synchronous method (e.g.
+/// `close()`) does not throw an exception immediately, but may throw it later
+/// asynchronously.\
+/// This can be useful in cases where it is known in advance that an error will
+/// (or may) be thrown, and this situation will be handled in a special way.
+Future<T> runAndDetach<T>(FutureOr<T> Function() action) {
+  final completer = Completer<T>();
+  runZonedGuarded(() async {
+    final result = await action();
+    completer.complete(result);
+  }, (error, stackTrace) {
+    if (!completer.isCompleted) {
+      completer.completeError(error, stackTrace);
+    }
+  });
+
+  return completer.future;
+}
+
+/// Performs the following actions:
+///
+/// - Adds a cancellation handler [onCancel]
+/// - Executes the [action] function
+/// - Removes a cancellation handler [onCancel]
+/// - Throws an [TaskCanceledError] exception if there was a cancellation
+/// request and no [TaskCanceledError] exception was thrown during the
+/// execution of the [action] function
+///
+/// The [onCancel] handler function should initiate the cancellation procedure
+/// which interrupts (or cancel) the execution of the [action] function.
+Future<void> runCancellable(
+  CancellationToken token,
+  void Function() onCancel,
+  FutureOr<void> Function() action,
+) async {
+  var isExceptionThrown = false;
+  final handler = token.addHandler(() {
+    onCancel();
+  });
+
+  try {
+    await action();
+  } catch (e) {
+    if (e is TaskCanceledError) {
+      isExceptionThrown = true;
+    }
+
+    rethrow;
+  } finally {
+    token.removerHandler(handler);
+    if (token.isCancelled && !isExceptionThrown) {
+      token.throwIfCancelled();
+    }
+  }
+}
 
 class CancellationToken {
   final Set<void Function()> _handlers = {};
@@ -57,30 +119,11 @@ class CancellationToken {
 
   // Removes the handler.\
   // The subscriber must call this method itself after the handler is no longer
-  //needed to free up memory.
+  // needed to free up memory.
   void removerHandler(void Function()? handler) {
     if (handler != null) {
       _handlers.remove(handler);
     }
-  }
-
-  /// Perform an asynchronous [action] with guard and cancellation handling.
-  @experimental
-  Future<void> runGuarded(
-    Future<void> Function() action, {
-    required FutureOr<void> Function() onCancel,
-  }) async {
-    await defer(() async {
-      try {
-        await onCancel();
-      } finally {
-        throwIfCancelled();
-      }
-    }, () async {
-      await runZonedGuarded(() async {
-        await action();
-      }, _handleErrors);
-    });
   }
 
   // Throw the exception [TaskCanceledError] if the token is in the `canceled`
@@ -88,14 +131,6 @@ class CancellationToken {
   void throwIfCancelled() {
     if (_isCancelled) {
       throw TaskCanceledError();
-    }
-  }
-
-  void _handleErrors(Object error, StackTrace stackTrace) {
-    if (error is TaskCanceledError) {
-      Error.throwWithStackTrace(error, stackTrace);
-    } else if (!_isCancelled) {
-      Error.throwWithStackTrace(error, stackTrace);
     }
   }
 }